cram/reader: Remove useless conversion

+++ noodles-cram/src/reader.rs
@@ -130,7 +130,7 @@ where
     pub(crate) fn read_container(&mut self) -> io::Result<Container> {
         let header = container::read_header(&mut self.inner)?;
 
-        let blocks_len = header.block_count() as usize;
+        let blocks_len = header.block_count();
         let mut blocks = Vec::with_capacity(blocks_len);
 
         for _ in 0..blocks_len {
